Fix to avoid Out of State Indications to apps
During the Out Of Service (OOS) bursts, an existing data connection is
reported as disconnected and the corresponding link properties are
removed, resulting in data connection loss.

We need to ignore OOS bursts. If the service is indeed lost, data call
disconnect indication will be sent by modem, else we will wait
optimistically for a quick reattach.

Add a system property to ignore out of service bursts.

Change-Id: I36f709b4cfc04a68542ec64c905f0850d2db231d
CRs-Fixed: 593413