MMS: Add functionality to store messages from operator
When receiving MMS from operators server number, it will save
in part table with both name and content location as null.

Add a check for content ID when making the PDU body, so as
to make sure either the name, file name, content location/ID
of the PDU part is not null. Otherwise, return ERROR.

Change-Id: Ie8067c4abc4e52d70d11b8c7b615d2664353ba5a
CRs-Fixed: 650563